Converted js files in expo-print to typescript
