--- packages/expo-print/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  get Print() {
-    return require('./src/Print');
-  },
-};
