+++ packages/expo-print/src/index.ts
@@ -1,0 +1,1 @@
+export { default as Print } from './Print';
