+++ packages/expo-print/tsconfig.json
@@ -1,0 +1,10 @@
+// @generated by expo-module-scripts
+{
+  // With TypeScript 3.2 this can become just "expo-module-scripts/tsconfig"
+  "extends": "../expo-module-scripts/tsconfig.base",
+  "compilerOptions": {
+    "outDir": "./build"
+  },
+  "include": ["./src"],
+  "exclude": ["**/__mocks__/*", "**/__tests__/*"]
+}
